Skip genome build validation for incremental upload
Inc. upload does not have meta_study.txt where genome build is
specified.
There is a ticket to read this info from API. We relax this check for
now to proceed with uploads. The loader gives and warning any way.
Although there is another ticket to make loader give an error in such
case.